[refactor]: URLSessionProtocol 수정에 따른 StubURLSession 변경
